Add custom class to email field in decorator
Since we're already using a decorator to unmask the email address it
makes sense to also add a custom class to the field. This makes it
easier to locate the field in the scraper class and means that the xpath
for finding the email fields only needs to be maintained in one place.