Also scrape Ministers page
Ministers can be appointed from outside the House, and automatically
become members, so we also need to scrape their page.

However, ministers who were already members have a page in both
sections, with a different ID on each, so we need to filter out any that
appear twice. (We assume here that no two members have the same name.)